fix(layout): fixes broken container max-width value
affects: @buildit/gravity-ui-web

The \`_layout.scss\` file in the "settings" layer was trying to reference a function defined later on
in the "tools" layer. This doesn't work and so the function's name was output in the CSS. Since it's
not valid CSS that value was broken, which in turn messed up other things. It has now been replaced
with a SASS var (and the expression has been simplied since the breakpoint widht and inset value now
use the same unit, so the calc() expression was no longer needed).