feat(column layout): adds mixin for full-width items
affects: @buildit/gravity-ui-web

The \`.grav-c-list-cards-basic\` component has an element modifier class that can stretch a single
card to span both columns. The CSS logic behind it might be useful to other column layouts, so this
change extracts and genericises it into a mixin that can be used in conjunction with any "no
danglies" column layout. It also updates the \`.grav-c-list-cards.basic\` class to use this mixin.