fix: adds missing width property to grav-l-container() mixin
affects: @buildit/gravity-ui-web

The \`grav-l-container\` mixin, that is used by the \`.grav-o-container\` class and also within the page
header and footer components, sets a \`max-width\` but does not set an explicit \`width\`. Normally this
is not an issue since block elements will naturally expand horizontally to fill the available space.
However, when used on direct descendents of \`<body>\`, which are laid out by flexbox 9for the sticky
footer behaviour), the elements shrink to fit the width of their content which causes them to then
appear centered on the page. This fix adds a \`width: 100%;\` to the mixin, which counteracts this
behaviour.