feat(layout): adds sticky-footer mixin and object class
affects: @buildit/gravity-ui-nunjucks, @buildit/gravity-ui-web

This change improves our sticky footer implementation, so that it is less likely to have unwanted
layout side-effects like spacing out top-level elements on short pages or altering the height of
`<main>`. It also wraps it up as a mixin, so that it is more easily re-usable. Finally, it adds a
`.grav-o-sticky-footer` class, which can be applied toroot containers that are not `<body>` - e.g.
the top-level roots you often find in SPAs as well as a `grav-o-undo-sticky-footer` class which
undoes the effect.