refactor: tightens up a linter exception
affects: @buildit/gravity-ui-web

There was an instance where Stylelint had to be intentionally disabled for one line, but it was done
in such a way that the linting rule was disabled for the remainder of that file. This fixes it, so
that the exception only applies to that one line.