feat: updates to work with ESLint 6.x
affects: @buildit/eslint-config

ESLint 6 handles the resolution of plugins and shared configs differently to version 5. As a
consequence, packages that use our shared config, and up till now had the required AirBnb shared
config installed as their own dev dependencies, break when they updgrade to ESLint 6.

The good news is that the new ESLint behaviour means we no longer need that config as a peer
dependency. Instead this shareable config can have its own dependency on it and the consumers no
longer need to worry about that.

They DO however still need to install eslint-plugin-import as a dependency of their own.

BREAKING CHANGE:
packages using this shared config need to upgrade to ESLint 6.x and remove the old
eslint-config-airbnb-base dependency.