fix(layouts): n-column layouts no longer cause horizontal scrolling
affects: @buildit/gravity-ui-web

This was an issue observed on the Buildit website, where some pages feature basic card lists, which
make use of our grav-l-column-layout() mixin. The negative left & right margins we previously used
extended beyond the width of the viewport at some screensizes. It appears that browsers treat the
right-hand negative margin as overflow in this case and thus add a horizontal scrollbar. The same
does not happen on the left side, so this fix changes the mixin to exclusively use left margins for
the column gaps.