fix(svg symbols): fixes missing title IDs
affects: @buildit/gravity-ui-web

Fixes a bug in our build process that caused the IDs for SVG symbols' <title> elements to be
omitted. This cuased an a11y issue because those IDs are needed in order for inlined instances of
those SVGs on a page to have text alternatives.

ISSUES CLOSED: #369